Rewrote tiered weights

+++ src/main/java/com/elenai/elenaidodge/capability/weight/IWeight.java
@@ -2,7 +2,7 @@ package com.elenai.elenaidodge.capability.weight;
 
 public interface IWeight {
 
-	public void set(int weight);
+	public void set(double weight);
 
-	public int getWeight();
+	public double getWeight();
 }
